Revert "chore: upstream the frame tree snapshot (#35902)"
This reverts commit 0876b99a4dbeb4ea101f37b3ba8a353a0113b077.
