+++ packages/protocol/src/channels.d.ts
@@ -2589,7 +2589,6 @@ export interface FrameChannel extends FrameEventTarget, Channel {
   addScriptTag(params: FrameAddScriptTagParams, metadata?: CallMetadata): Promise<FrameAddScriptTagResult>;
   addStyleTag(params: FrameAddStyleTagParams, metadata?: CallMetadata): Promise<FrameAddStyleTagResult>;
   ariaSnapshot(params: FrameAriaSnapshotParams, metadata?: CallMetadata): Promise<FrameAriaSnapshotResult>;
-  snapshotForAI(params: FrameSnapshotForAIParams, metadata?: CallMetadata): Promise<FrameSnapshotForAIResult>;
   blur(params: FrameBlurParams, metadata?: CallMetadata): Promise<FrameBlurResult>;
   check(params: FrameCheckParams, metadata?: CallMetadata): Promise<FrameCheckResult>;
   click(params: FrameClickParams, metadata?: CallMetadata): Promise<FrameClickResult>;
@@ -2696,21 +2695,14 @@ export type FrameAddStyleTagResult = {
 };
 export type FrameAriaSnapshotParams = {
   selector: string,
+  forAI?: boolean,
   timeout?: number,
 };
 export type FrameAriaSnapshotOptions = {
+  forAI?: boolean,
   timeout?: number,
 };
 export type FrameAriaSnapshotResult = {
-  snapshot: string,
-};
-export type FrameSnapshotForAIParams = {
-  selector: string,
-};
-export type FrameSnapshotForAIOptions = {
-
-};
-export type FrameSnapshotForAIResult = {
   snapshot: string,
 };
 export type FrameBlurParams = {
