+++ tests/page/page-aria-snapshot-ai.spec.ts
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
+import type { FrameLocator, Page } from '@playwright/test';
 import { test as it, expect } from './pageTest';
 
-function snapshotForAI(page: any): Promise<string> {
-  return page.locator('body')._snapshotForAI();
-}
+const forAI = { _forAI: true } as any;
 
 it('should generate refs', async ({ page }) => {
   await page.setContent(`
@@ -27,7 +26,7 @@ it('should generate refs', async ({ page }) => {
     <button>Three</button>
   `);
 
-  const snapshot1 = await snapshotForAI(page);
+  const snapshot1 = await page.locator('body').ariaSnapshot(forAI);
   expect(snapshot1).toContainYaml(`
     - generic [ref=e1]:
       - button "One" [ref=e2]
@@ -42,7 +41,7 @@ it('should generate refs', async ({ page }) => {
     e.textContent = 'Not Two';
   });
 
-  const snapshot2 = await snapshotForAI(page);
+  const snapshot2 = await page.locator('body').ariaSnapshot(forAI);
   expect(snapshot2).toContainYaml(`
     - generic [ref=e1]:
       - button "One" [ref=e2]
@@ -57,17 +56,37 @@ it('should list iframes', async ({ page }) => {
     <iframe name="foo" src="data:text/html,<h1>World</h1>">
   `);
 
-  const snapshot1 = await snapshotForAI(page);
+  const snapshot1 = await page.locator('body').ariaSnapshot(forAI);
   expect(snapshot1).toContain('- iframe');
 
   const frameSnapshot = await page.frameLocator(`iframe`).locator('body').ariaSnapshot();
   expect(frameSnapshot).toEqual('- heading "World" [level=1]');
 });
 
-it('should stitch all frame snapshots', async ({ page, server }) => {
+it('ref mode can be used to stitch all frame snapshots', async ({ page, server }) => {
   await page.goto(server.PREFIX + '/frames/nested-frames.html');
-  const snapshot = await snapshotForAI(page);
-  expect(snapshot).toContainYaml(`
+
+  async function allFrameSnapshot(frame: Page | FrameLocator): Promise<string> {
+    const snapshot = await frame.locator('body').ariaSnapshot(forAI);
+    const lines = snapshot.split('\n');
+    const result = [];
+    for (const line of lines) {
+      const match = line.match(/^(\s*)- iframe \[ref=(.*)\]/);
+      if (!match) {
+        result.push(line);
+        continue;
+      }
+
+      const leadingSpace = match[1];
+      const ref = match[2];
+      const childFrame = frame.frameLocator(`aria-ref=${ref}`);
+      const childSnapshot = await allFrameSnapshot(childFrame);
+      result.push(line + ':', childSnapshot.split('\n').map(l => leadingSpace + '  ' + l).join('\n'));
+    }
+    return result.join('\n');
+  }
+
+  expect(await allFrameSnapshot(page)).toContainYaml(`
     - generic [ref=e1]:
       - iframe [ref=e2]:
         - generic [ref=e1]:
@@ -87,7 +106,7 @@ it('should not generate refs for hidden elements', async ({ page }) => {
     <button>Three</button>
   `);
 
-  const snapshot = await snapshotForAI(page);
+  const snapshot = await page.locator('body').ariaSnapshot(forAI);
   expect(snapshot).toContainYaml(`
     - generic [ref=e1]:
       - button "One" [ref=e2]
@@ -119,7 +138,7 @@ it('should not generate refs for elements with pointer-events:none', async ({ pa
     </div>
   `);
 
-  const snapshot = await snapshotForAI(page);
+  const snapshot = await page.locator('body').ariaSnapshot(forAI);
   expect(snapshot).toContainYaml(`
     - generic [ref=e1]:
       - button "no-ref"
@@ -163,7 +182,7 @@ it('emit generic roles for nodes w/o roles', async ({ page }) => {
     </div>
   `);
 
-  const snapshot = await snapshotForAI(page);
+  const snapshot = await page.locator('body').ariaSnapshot(forAI);
 
   expect(snapshot).toContainYaml(`
     - generic [ref=e2]:
@@ -193,7 +212,7 @@ it('should collapse generic nodes', async ({ page }) => {
     </div>
   `);
 
-  const snapshot = await snapshotForAI(page);
+  const snapshot = await page.locator('body').ariaSnapshot(forAI);
   expect(snapshot).toContainYaml(`
     - button \"Button\" [ref=e5]
   `);
@@ -204,7 +223,7 @@ it('should include cursor pointer hint', async ({ page }) => {
     <button style="cursor: pointer">Button</button>
   `);
 
-  const snapshot = await snapshotForAI(page);
+  const snapshot = await page.locator('body').ariaSnapshot(forAI);
   expect(snapshot).toContainYaml(`
     - button \"Button\" [ref=e2] [cursor=pointer]
   `);
